[#13] Minor source code improvements (Part 42: Resolving compiler warnings C4244, C4800, and C4018)

+++ Main/src/SpectrumDos_FileManagerBase.cpp
@@ -25,7 +25,7 @@
 		const PDirectoryTraversal pdt=DOS->BeginDirectoryTraversal();
 			if (!pdt) return nullptr;
 			::memcpy( tmpDirEntry, file, pdt->entrySize );
-			const BYTE nameCharsMax=pdt->nameCharsMax;
+			const WORD nameCharsMax=pdt->nameCharsMax;
 		DOS->EndDirectoryTraversal(pdt);
 		for( BYTE copyNumber=1; copyNumber; copyNumber++ ){
 			// . composing the Name for the File copy
